+++ rich/default_styles.py
@@ -116,6 +116,8 @@ DEFAULT_STYLES: Dict[str, Style] = {
     "progress.percentage": Style(color="magenta"),
     "progress.remaining": Style(color="cyan"),
     "progress.data.speed": Style(color="red"),
+    "progress.spinner": Style(color="green"),
+    "status.spinner": Style(color="green"),
 }
 
 MARKDOWN_STYLES = {