+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added rich.live https://github.com/willmcgugan/rich/pull/382
 - Added algin parameter to Rule and Console.rule
+- Added rich.Status class and Console.status
+- Added getitem to Text
+- Added style parameter to Console.log
+
+### Changed
+
+- Table.add_row style argument now applies to entire line and not just cells
+- Added end_section parameter to Table.add_row to force a line underneath row
 
 ## [9.3.0] - 2020-12-1